requirements.txt
fastapi
google-generativeai
pydantic
uvicorn
python-dotenv
